+++ config/splat.us.dra.yaml
@@ -218,7 +218,7 @@ segments:
       # - [0x41D84, .rodata, 7E4BC] # EntityPlayerPinkEffect
       # - [0x41DA4, .rodata, 7E4BC] # EntityPlayerDissolves
       # - [0x41DBC, .rodata, 7E4BC] # EntityMist
-      - [0x41DD4, .rodata, 843B0] # EntityTeleport
+      # - [0x41DD4, .rodata, 7E4BC] # EntityTeleport
       - [0x41E34, .rodata, 84B88] # EntitySubwpnCrashCross
       # - [0x41E48, .rodata, 84B88] # EntityHellfireHandler
       # - [0x41E7C, .rodata, 84B88] # func_80128C2C
@@ -268,7 +268,6 @@ segments:
       - [0x75F54, c, 75F54]
       - [0x7879C, c, 7879C]
       - [0x7E4BC, c, 7E4BC]
-      - [0x843B0, c, 843B0]
       - [0x84B88, c, 84B88]
       - [0x8A0A4, c, 8A0A4]
       - [0x8BEF8, c, 8BEF8]
