+++ src/dra/7E4BC.c
@@ -2359,3 +2359,262 @@ void func_80124164(
         prim->b0 = prim->b2 = D_800AE290 * colorIntensity / 0x100;
     }
 }
+
+// BSS
+extern Point16 D_8013839C[32];
+
+// teleport effect like when using library card (ID 0x42)
+void EntityTeleport(Entity* self) {
+    Primitive* prim;
+    s32 selfUnk7C;
+    s32 selfUnk80;
+    s32 yVar;
+    s32 xVar;
+    s32 upperParams;
+    s32 i;
+    s32 result;
+
+    bool wasCase3 = false;
+    bool var_s5 = false;
+
+    upperParams = self->params & 0xFE00;
+    switch (self->step) {
+    case 0:
+        self->primIndex = AllocPrimitives(PRIM_GT4, LEN(D_8013839C) + 4);
+        if (self->primIndex == -1) {
+            return;
+        }
+        self->flags = FLAG_POS_CAMERA_LOCKED | FLAG_KEEP_ALIVE_OFFCAMERA |
+                      FLAG_HAS_PRIMS | FLAG_UNK_10000;
+        prim = &g_PrimBuf[self->primIndex];
+        for (i = 0; i < 2; i++) {
+            prim->x0 = 0xC0 * i;
+            prim->g0 = 0;
+            prim->b0 = 0;
+            prim->r0 = 0;
+            prim->y0 = 0;
+            prim->u0 = 0xC0;
+            prim->v0 = 0xF0;
+            prim->type = PRIM_TILE;
+            prim->priority = 0x1FD;
+            prim->drawMode = DRAW_TPAGE2 | DRAW_TPAGE | DRAW_HIDE | DRAW_TRANSP;
+            prim = prim->next;
+        }
+        for (i = 0; i < 2; i++) {
+            prim->type = PRIM_G4;
+            prim->priority = 0x1F8;
+            prim->drawMode = DRAW_TPAGE2 | DRAW_TPAGE | DRAW_TRANSP;
+            prim = prim->next;
+        }
+        for (i = 0; i < LEN(D_8013839C); i++) {
+            xVar = PLAYER.posX.i.hi + (rand() % 28) - 14;
+            yVar = rand();
+            yVar = 0xE0 - (yVar & 0x3F);
+            D_8013839C[i].x = xVar;
+            D_8013839C[i].y = yVar;
+            prim->clut = 0x1BA;
+            prim->tpage = 0x1A;
+            prim->b0 = 0;
+            prim->b1 = 0;
+            prim->g0 = 0;
+            prim->g1 = (rand() & 0x1F) + 1;
+            prim->priority = 0x1F0;
+            prim->drawMode = DRAW_HIDE;
+            prim->g2 = 0;
+            prim = prim->next;
+        }
+        self->ext.teleport.unk7C = 0;
+        self->ext.teleport.unk80 = 0x10;
+        self->ext.teleport.unk88 = 0x80;
+        if (self->params & 0x100) {
+            var_s5 = true;
+            self->ext.teleport.unk7C = 0x10;
+            self->ext.teleport.unk80 = 0x100;
+            self->ext.teleport.unk88 = 0x80;
+            self->ext.teleport.unk90 = 0xFF;
+            self->step = 0x14;
+            PlaySfx(SFX_UNK_8BB);
+        } else {
+            self->ext.teleport.unk7C = 1;
+            self->ext.teleport.unk90 = 0;
+            self->ext.teleport.unk80 = 0x10;
+            self->ext.teleport.unk88 = 0x80;
+            self->step = 1;
+            PlaySfx(SFX_TELEPORT_BANG_A);
+            PlaySfx(NA_SE_PL_TELEPORT);
+        }
+        break;
+    case 1:
+        self->ext.teleport.unk80 += 0x20;
+        if (self->ext.teleport.unk80 >= 0x101) {
+            self->step++;
+        }
+        break;
+    case 2:
+        if (++self->ext.teleport.unk7C >= 0x10) {
+            self->ext.teleport.unk7C = 0x10;
+            self->ext.teleport.unk84 = 0x80;
+            self->step++;
+        }
+        break;
+    case 3:
+        wasCase3 = true;
+        self->ext.teleport.unk88 += 4;
+        if (self->ext.teleport.unk88 >= 0x100) {
+            self->ext.teleport.unk88 = 0x100;
+        }
+        if (--self->ext.teleport.unk84 == 0) {
+            PLAYER.palette = 0x810D;
+            self->step++;
+        }
+        break;
+    case 4:
+        PLAYER.palette = 0x810D;
+        if (--self->ext.teleport.unk7C <= 0) {
+            self->ext.teleport.unk7C = 0;
+            self->step++;
+        }
+        break;
+    case 5:
+        var_s5 = true;
+        PLAYER.palette = 0x810D;
+        self->ext.teleport.unk90 += 4;
+        if (self->ext.teleport.unk90 >= 0x100) {
+            self->ext.teleport.unk90 = 0xFF;
+            self->ext.teleport.unk84 = 0x20;
+            self->step++;
+        }
+        break;
+    case 6:
+        var_s5 = true;
+        PLAYER.palette = 0x810D;
+        if (--self->ext.teleport.unk84 == 0) {
+            self->ext.teleport.unk90 = 0;
+            if (upperParams == 0) {
+                D_80097C98 = 6;
+            }
+            if (upperParams == 0x200) {
+                D_80097C98 = 4;
+            }
+            if (upperParams == 0x400) {
+                D_80097C98 = 5;
+            }
+        }
+        break;
+    case 20:
+        var_s5 = true;
+        self->ext.teleport.unk90 = 0xFF;
+        self->ext.teleport.unk84 = 0x20;
+        self->step++;
+        break;
+    case 21:
+        var_s5 = true;
+        if (--self->ext.teleport.unk84 == 0) {
+            self->step++;
+        }
+        break;
+    case 22:
+        var_s5 = true;
+        self->ext.teleport.unk90 -= 4;
+        if (self->ext.teleport.unk90 <= 0) {
+            self->ext.teleport.unk90 = 0;
+            self->step++;
+        }
+        break;
+    case 23:
+        if (--self->ext.teleport.unk7C < 2) {
+            self->ext.teleport.unk7C = 0;
+            self->ext.teleport.unk84 = 4;
+            self->step++;
+            g_Player.unk1C = 1;
+            PlaySfx(SFX_TELEPORT_BANG_B);
+            DestroyEntity(self);
+            return;
+        }
+        break;
+    }
+    selfUnk7C = self->ext.teleport.unk7C;
+    selfUnk80 = self->ext.teleport.unk80;
+    self->posX.i.hi = PLAYER.posX.i.hi;
+    self->posY.i.hi = PLAYER.posY.i.hi;
+    prim = &g_PrimBuf[self->primIndex];
+    xVar = PLAYER.posX.i.hi;
+    yVar = PLAYER.posY.i.hi;
+    for (i = 0; i < 2; i++) {
+        prim->r0 = prim->b0 = prim->g0 = self->ext.teleport.unk90;
+        prim->drawMode |= DRAW_HIDE;
+        if (var_s5) {
+            prim->drawMode &= ~DRAW_HIDE;
+        }
+        prim = prim->next;
+    }
+    prim->x1 = prim->x3 = xVar;
+    prim->x0 = prim->x2 = xVar - selfUnk7C;
+    func_80124164(prim, self->ext.teleport.unk88, yVar, selfUnk80, upperParams);
+    prim = prim->next;
+    prim->x1 = prim->x3 = xVar;
+    prim->x0 = prim->x2 = xVar + selfUnk7C;
+    func_80124164(prim, self->ext.teleport.unk88, yVar, selfUnk80, upperParams);
+    prim = prim->next;
+    if (wasCase3) {
+        for (i = 0; i < LEN(D_8013839C); i++) {
+            switch (prim->g0) {
+            case 0:
+                if (--prim->g1 == 0) {
+                    prim->g0++;
+                }
+                break;
+            case 1:
+                result = func_80119E78(prim, D_8013839C[i].x, D_8013839C[i].y);
+                D_8013839C[i].y -= 16;
+                if (result < 0) {
+                    prim->drawMode |= DRAW_HIDE;
+                    prim->g0++;
+                } else {
+                    prim->drawMode &= ~DRAW_HIDE;
+                }
+                break;
+            }
+            prim = prim->next;
+        }
+        return;
+    }
+
+    // Potential bug? Should probably be doing prim = prim->next, right?
+    for (i = 0; i < 32; i++) {
+        prim->drawMode |= DRAW_HIDE;
+    }
+}
+
+// Entity #66. Made by final blueprint, #122. Created in PlayerStepStand.
+void EntityPlayerSleepZ(Entity* self) {
+    if (PLAYER.step != Player_Stand || PLAYER.step_s != Player_Stand_ChairSit) {
+        DestroyEntity(self);
+        return;
+    }
+
+    switch (self->step) {
+    case 0:
+        self->animSet = ANIMSET_DRA(17);
+        self->velocityY = FIX(-0.375);
+        SetSpeedX(FIX(0.25));
+        self->unk5A = 0x50;
+        self->palette = PAL_OVL(0x19F);
+        self->anim = D_800AE294;
+
+        self->posY.i.hi -= 16;
+        self->posX.val += self->velocityX << 5;
+        self->flags = FLAG_UNK_100000;
+        self->facingLeft = 0;
+        self->step++;
+        break;
+
+    case 1:
+        self->posX.val += self->velocityX;
+        self->posY.val += self->velocityY;
+        if (self->animFrameDuration < 0) {
+            DestroyEntity(self);
+        }
+        break;
+    }
+}
